refactor: 매물 클릭시 깜빡거리는 현상 제거 및 Map.tsx 일부 리팩토링
