+++ src/routers/MapRouter.tsx
@@ -4,7 +4,7 @@ import Map from '../routes/map/Map';
 import MapSearch from '../routes/map/MapSearch';
 import MainHeader from '../components/common/header/MainHeader';
 import SubHeader from '@components/common/header/SubHeader';
-import MapPropertyLoan from '@/routes/map/MapPropertyLoan';
+import MapPropertyLoan from '@components/map/MapPropertyLoan';
 
 export const mapRoutes: RouteObject[] = [
   {
